requirements.txt
allure-behave==2.13.2
allure-combine==1.0.11
allure-python-commons==2.13.2
Appium-Python-Client==2.11.1
attrs==23.2.0
beautifulsoup4==4.12.3
behave==1.2.6
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==5.11.2
h11==0.14.0
httplib2==0.22.0
idna==3.7
iniconfig==2.0.0
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.0
parse==1.20.1
parse-type==0.6.2
pillow==10.3.0
pluggy==1.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
PySocks==1.7.1
pytest==7.4.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.11.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.0
trio-websocket==0.11.1
urllib3==1.26.6
wsproto==1.2.0
pytz==2024.1